make prosemirror work in shadow dom
shadowRoot is a DOM Node and it is the standard (https://twitter.com/MaoStevemao/status/767864034049732608). We keep a reference of `root` in prosemirror instance.

I tried to do a global search in src and compare `document`/`window` with https://developer.mozilla.org/en-US/docs/Web/API/ShadowRoot, to see if we could use `pm.root` DOM Node instead. There might be something missing either in the MDN docs or in the source code.